@@ -4,7 +4,7 @@ project(SuperPool)
 
 set(LIBRARY_OUTPUT_DIRECTORY ${PROJECT_SOURCE_DIR}/lib)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -std=c++14" )
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g -std=c++17" )
 include_directories(${PROJECT_SOURCE_DIR}/include)
 
 aux_source_directory(./src SRC_LIST)
